requirements.txt
ale_py==0.10.1
chex==0.1.86
cloudpickle==3.1.0
colored==2.2.4
diffusers==0.29.2
dm_control==1.0.20
dm_env==1.6
gym==0.26.2
gymnasium==0.29.1
hydra-core==1.3.2
imageio[ffmpeg]==2.34.2
jax[cuda12]==0.4.28
labmaze==1.0.6
matplotlib==3.9.2
mlflow==2.17.0
msgpack_python==0.5.6
mujoco==3.1.6
numpy==1.24.3
omegaconf==2.3.0
optax==0.2.2
pandas==2.2.3
psutil==6.0.0
pytest==8.3.3
pyzmq==26.2.0
rich==13.7.1
ruamel.yaml==0.18.6
tensorboard==2.16.0
tensorflow==2.16.2
tensorflow-probability==0.24.0
tf-keras==2.16.0
termcolor==2.5.0
torch==2.3.1 -f https://download.pytorch.org/whl/cu121/torch_stable.html
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
tyro==0.8.5
wandb[media]==0.18.3
